Fix non-sensical test
This test was actually failing for a while because a Promise was not properly
awaited, which masked the failure. The test itself does not make a whole lot of
sense, because it was checking whether the updatePackage command actually adds
new packages to a project, which should not really happen.